% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{build_deps}
\alias{build_deps}
\title{Find Build-Time Package Dependencies}
\usage{
build_deps(path = ".")
}
\arguments{
\item{path}{Path to the package root.}
}
\value{
Character vector, the names of the packages that are
build dependencies.
}
\description{
Most R package dependencies are run-time dependencies: functions within
a package refer to functions or other objects within another package.
These references are resolved at runtime, and for evaluating the code
of a package (i.e. installing it), the dependencies are not actually
needed. This package tries to work out the build-time and run-time
dependencies of a package, by trying to evaluate the package code both
without and with each dependency.
}
\details{
The current algorithms is this:
\itemize{
\item If a dependency is linked to (i.e. its type is \code{LinkingTo}), this is
a build-time dependency, and we install it.
\item We try to evaluate the package code with only the \code{LinkingTo}
dependencies available. This should already run without errors for most
packages, and if it indeed does, then there are no additional
dependencies.
\item Otherwise we install all dependencies first, and then employ a
"leave one out" strategy for the non \code{LinkingTo} packages: for each
imported or depended package, we omit this single package and try to
evaluate the package code.
}
}

