% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{del_pkg_from_depgraph}
\alias{del_pkg_from_depgraph}
\title{Remove a package (or more) from the dependency graph}
\usage{
del_pkg_from_depgraph(graph, pkg)
}
\arguments{
\item{graph}{Dependency graph.}

\item{pkg}{Package(s) to remove.}
}
\value{
The cleaned up dependency graph.
}
\description{
Remove a package (or more) from the dependency graph
}
\keyword{internal}
