% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{filter_namespace}
\alias{filter_namespace}
\title{Remove some imports from the namespace}
\usage{
filter_namespace(ns, without)
}
\arguments{
\item{ns}{The \code{namespace} object, as read by the \code{parseNamespaceFile}
function.}

\item{without}{Character vector of package names to omit from the
imports.}
}
\value{
A \code{namespace} object, after the \code{without} packages were
removed from the imports.
}
\description{
Remove some imports from the namespace
}
\keyword{internal}
