% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{make_depgraph}
\alias{make_depgraph}
\title{Create a package dependency graph}
\usage{
make_depgraph(pkgfiles)
}
\arguments{
\item{pkgfiles}{The names of the package files. We assume that all
(non-base) recursive dependencies are included here.}
}
\value{
The dependency graph, a named list, an out-adjacency list.
}
\description{
This only contains all recursive dependencies of the package under study,
since it might happen that a dependency depends on another dependency
indirectly, and we need to know about this.
}
\keyword{internal}
