% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{next_pkg_from_depgraph}
\alias{next_pkg_from_depgraph}
\title{Yield the next package to try from the depgraph}
\usage{
next_pkg_from_depgraph(graph)
}
\arguments{
\item{graph}{The dependency graph.}
}
\value{
Character scalar, the name of the next package to try.
}
\description{
Yield the next package to try from the depgraph
}
\keyword{internal}
