% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ builddeps }
\alias{ builddeps }
\alias{ builddeps }
\title{ Find Build-Time Package Dependencies }
\description{
Most R package dependencies are run-time dependencies: functions within a
  package refer to functions or other objects within another package. These
  references are resolved at runtime, and for evaluating the code of a package
  (i.e. installing it), the dependencies are not actually needed. This package
  tries to work out the build-time and run-time dependencies of a package, by
  trying to evaluate the package code both without and with each dependency.
}
