% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try-load.R
\name{try_load}
\alias{try_load}
\title{Try loading a package from a directory}
\usage{
try_load(path, libpath, pkg)
}
\arguments{
\item{path}{Path to the package directory.}

\item{libpath}{The library path to set.}
}
\value{
Flag (logical scalar), whether the package loading was
successful or not.
}
\description{
It tries the loading in another R session.
}
\keyword{internal}
