% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{write_namespace}
\alias{write_namespace}
\title{Write out a \code{NAMESPACE} file}
\usage{
write_namespace(ns, path)
}
\arguments{
\item{ns}{The \code{namespace} object, as read by the \code{parseNamespaceFile}
function.}

\item{path}{Path to the namespace file.}
}
\description{
Note that this does \strong{not} reproduce the original namespace,
e.g. R code for conditional imports is lost. For \code{builddeps} this is OK,
because we parse \code{NAMESPACE} and use the modified version on the
same system, so they should evaluate the same way.
}
\keyword{internal}
